@@ -86,36 +91,34 @@
     "jest-environment-node": "^29.7.0",
     "jest-junit": "^16.0.0",
     "jest-location-mock": "^2.0.0",
-    "module-alias": "^2.2.2",
-    "nanoid": "^3.1.31",
-    "npm-run-all": "^4.1.5",
+    "memfs": "^4.9.2",
+    "module-alias": "^2.2.3",
+    "nanoid": "^3.3.8",
+    "npm-run-all2": "^7.0.2",
     "prettier": "^3.3.0",
-    "puppeteer": "^22.10.0",
-    "react-refresh": "^0.14.0",
-    "semver": "^7.5.2",
+    "puppeteer": "^24.4.0",
+    "react-refresh": "^0.17.0",
     "sourcemap-validator": "^2.1.0",
-    "type-fest": "^4.0.0",
+    "terser-webpack-plugin": "^5.3.10",
+    "type-fest": "^4.18.3",
     "typescript": "~5.4.5",
-    "webpack": "^5.76.0",
+    "webpack": "^5.94.0",
     "webpack-cli": "^5.1.4",
-    "webpack-cli-v4": "npm:webpack-cli@4.x",
     "webpack-dev-server": "^5.0.4",
-    "webpack-dev-server-v3": "npm:webpack-dev-server@3.x",
-    "webpack-dev-server-v4": "npm:webpack-dev-server@4.x",
-    "webpack-hot-middleware": "^2.25.0",
-    "webpack-plugin-serve": "^1.4.1",
-    "webpack-v4": "npm:webpack@4.x",
+    "webpack-dev-server-v4": "npm:webpack-dev-server@^4.8.0",
+    "webpack-hot-middleware": "^2.26.1",
+    "webpack-plugin-serve": "^1.6.0",
     "yn": "^4.0.0"
   },
   "peerDependencies": {
-    "@types/webpack": "4.x || 5.x",
+    "@types/webpack": "5.x",
     "react-refresh": ">=0.10.0 <1.0.0",
     "sockjs-client": "^1.4.0",
     "type-fest": ">=0.17.0 <5.0.0",
-    "webpack": ">=4.43.0 <6.0.0",
-    "webpack-dev-server": "3.x || 4.x || 5.x",
+    "webpack": "^5.0.0",
+    "webpack-dev-server": "^4.8.0 || 5.x",
     "webpack-hot-middleware": "2.x",
-    "webpack-plugin-serve": "0.x || 1.x"
+    "webpack-plugin-serve": "1.x"
   },
   "peerDependenciesMeta": {
     "@types/webpack": {
@@ -138,10 +141,11 @@
     }
   },
   "resolutions": {
-    "type-fest": "^4.0.0"
+    "memfs": "^4.0.0",
+    "rimraf": "^5.0.0",
+    "type-fest": "^4.20.0"
   },
   "engines": {
-    "node": ">= 10.13"
-  },
-  "packageManager": "yarn@1.22.22+sha1.ac34549e6aa8e7ead463a7407e1c7390f61a6610"
+    "node": ">=18.12"
+  }
 }
