@@ -56,44 +54,57 @@
   ],
   "author": "Evgeny Poberezkin",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/ajv-validator/ajv/issues"
-  },
-  "homepage": "https://github.com/ajv-validator/ajv",
-  "tonicExampleFilename": ".tonic_example.js",
+  "bugs": "https://github.com/ajv-validator/ajv/issues",
+  "homepage": "https://ajv.js.org",
+  "runkitExampleFilename": ".runkit_example.js",
   "dependencies": {
-    "fast-deep-equal": "^3.1.1",
-    "fast-json-stable-stringify": "^2.0.0",
-    "json-schema-traverse": "^0.4.1",
-    "uri-js": "^4.2.2"
+    "fast-deep-equal": "^3.1.3",
+    "fast-uri": "^3.0.1",
+    "json-schema-traverse": "^1.0.0",
+    "require-from-string": "^2.0.2"
   },
   "devDependencies": {
-    "ajv-async": "^1.0.0",
-    "bluebird": "^3.5.3",
-    "brfs": "^2.0.0",
-    "browserify": "^16.2.0",
-    "chai": "^4.0.1",
-    "coveralls": "^3.0.1",
-    "del-cli": "^3.0.0",
-    "dot": "^1.0.3",
-    "eslint": "^7.3.1",
-    "gh-pages-generator": "^0.2.3",
-    "glob": "^7.0.0",
-    "if-node-version": "^1.0.0",
-    "js-beautify": "^1.7.3",
-    "jshint": "^2.10.2",
+    "@ajv-validator/config": "^0.5.0",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-typescript": "^11.1.6",
+    "@types/chai": "^4.3.11",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "^20.11.30",
+    "@types/require-from-string": "^1.2.3",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
+    "@typescript-eslint/parser": "^7.3.1",
+    "ajv-formats": "^3.0.1",
+    "browserify": "^17.0.0",
+    "chai": "^4.4.1",
+    "cross-env": "^7.0.3",
+    "dayjs": "^1.11.10",
+    "dayjs-plugin-utc": "^0.1.2",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "glob": "^10.3.10",
+    "husky": "^9.0.11",
+    "if-node-version": "^1.1.1",
+    "jimp": "^0.22.10",
+    "js-beautify": "^1.15.1",
     "json-schema-test": "^2.0.0",
-    "karma": "^5.0.0",
-    "karma-chrome-launcher": "^3.0.0",
-    "karma-mocha": "^2.0.0",
-    "karma-sauce-launcher": "^4.1.3",
-    "mocha": "^8.0.1",
-    "nyc": "^15.0.0",
-    "pre-commit": "^1.1.1",
-    "require-globify": "^1.3.0",
-    "typescript": "^3.9.5",
-    "uglify-js": "^3.6.9",
-    "watch": "^1.0.0"
+    "karma": "^6.4.2",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-mocha": "^2.0.1",
+    "lint-staged": "^15.2.2",
+    "mocha": "^10.3.0",
+    "module-from-string": "^3.3.0",
+    "node-fetch": "^3.3.2",
+    "nyc": "^15.1.0",
+    "prettier": "3.0.3",
+    "re2": "^1.20.9",
+    "rollup": "^2.79.1",
+    "rollup-plugin-terser": "^7.0.2",
+    "ts-node": "^10.9.2",
+    "tsify": "^5.0.4",
+    "typescript": "5.3.3",
+    "uri-js": "^4.4.1"
   },
   "collective": {
     "type": "opencollective",
