@@ -1,16 +1,21 @@
 {
 	"name": "p-retry",
-	"version": "4.6.2",
+	"version": "6.2.1",
 	"description": "Retry a promise-returning or async function",
 	"license": "MIT",
 	"repository": "sindresorhus/p-retry",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
+	"types": "./index.d.ts",
+	"sideEffects": false,
 	"engines": {
-		"node": ">=8"
+		"node": ">=16.17"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -39,13 +44,14 @@
 		"bluebird"
 	],
 	"dependencies": {
-		"@types/retry": "0.12.0",
+		"@types/retry": "0.12.2",
+		"is-network-error": "^1.0.0",
 		"retry": "^0.13.1"
 	},
 	"devDependencies": {
-		"ava": "^2.4.0",
-		"delay": "^4.1.0",
-		"tsd": "^0.10.0",
-		"xo": "^0.25.3"
+		"ava": "^5.3.1",
+		"delay": "^6.0.0",
+		"tsd": "^0.28.1",
+		"xo": "^0.56.0"
 	}
 }
