@@ -1,6 +1,6 @@
 {
   "name": "react-router-dom",
-  "version": "7.4.0",
+  "version": "7.6.2",
   "description": "Declarative routing for React web applications",
   "keywords": [
     "react",
@@ -54,11 +54,11 @@
     }
   },
   "dependencies": {
-    "react-router": "7.4.0"
+    "react-router": "7.6.2"
   },
   "devDependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "tsup": "^8.3.0",
     "typescript": "^5.1.6",
     "wireit": "0.14.9"
