@@ -1,6 +1,6 @@
 {
   "name": "react-router",
-  "version": "7.4.0",
+  "version": "7.6.2",
   "description": "Declarative routing for React",
   "keywords": [
     "react",
@@ -37,12 +37,15 @@
         "default": "./dist/development/index.js"
       }
     },
-    "./route-module": {
+    "./internal": {
+      "node": {
+        "types": "./dist/development/lib/types/internal.d.ts"
+      },
       "import": {
-        "types": "./dist/development/lib/types/route-module.d.mts"
+        "types": "./dist/development/lib/types/internal.d.mts"
       },
       "default": {
-        "types": "./dist/development/lib/types/route-module.d.ts"
+        "types": "./dist/development/lib/types/index.d.ts"
       }
     },
     "./dom": {
@@ -77,15 +80,13 @@
     }
   },
   "dependencies": {
-    "@types/cookie": "^0.6.0",
     "cookie": "^1.0.1",
-    "set-cookie-parser": "^2.6.0",
-    "turbo-stream": "2.4.0"
+    "set-cookie-parser": "^2.6.0"
   },
   "devDependencies": {
     "@types/set-cookie-parser": "^2.4.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "rimraf": "^6.0.1",
     "tsup": "^8.3.0",
     "typescript": "^5.1.6",
