@@ -1,30 +1,36 @@
- import React from "react";
- import IMAGES from "../data/images.js";
- import ImageShadow from "react-image-shadow";
- import "react-image-shadow/assets/index.css";
+import React, { useState } from "react";
+import IMAGES from "../data/images.js";
+import "../css/TechnologiesFuturistic.css";
 
- const Technologies = () => {
-    return(
-        <div> 
-          <h2 className="tech-heading">Technologies & Languages That Drive My Development Journey</h2>
-          <h3 className="tech-heading">Here’s What’s Under the Hood 🛠️</h3>
-       {IMAGES.map(imageObject => { //we are moving through the map function
-        return <Image key= {imageObject.id} src={imageObject.image} desc={imageObject.description}/>
-       })}
-        </div>
-    )
- }
- const Image = ({src , desc}) => { // we need to get them as input parameters , deconstruct way 
-    // or inside (props) and down write src,desc WAY2 
-    // we use a or <a/> element to be clickable 
-    // and we get the paramaters from the other component 
-    return (
-        <div className="technologies">
-     <a target="_blank" rel="" href={src}>
-        <ImageShadow className ="image-shadow"src={src}/></a>
-        <p>{desc}</p>
-        </div>
-    )
- }
+const Technologies = () => {
+  // Track which ball is animating
+  const [active, setActive] = useState(null);
 
- export default Technologies;
+  // Handle click: trigger animation for this ball
+  const handleClick = (id) => {
+    setActive(id);
+    setTimeout(() => setActive(null), 600); // Animation duration
+  };
+
+  return (
+    <div>
+      <h2 className="tech-heading">Technologies & Languages That Drive My Development Journey</h2>
+      <h3 className="tech-subheading">Click a Ball!</h3>
+      <div className="svg-ball-grid">
+        {IMAGES.map((img) => (
+          <div
+            key={img.id}
+            className={`svg-ball${active === img.id ? " animate" : ""}`}
+            onClick={() => handleClick(img.id)}
+            title={img.title}
+          >
+            <img src={img.image} alt={img.title} className="svg-ball-img" />
+            <div className="svg-ball-label">{img.title}</div>
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+};
+
+export default Technologies;
