@@ -1,4 +1,5 @@
 import React, { Component } from "react";
+import "../css/Title.css";
 
 const TITLES = [
     "Software Engineer", 
@@ -12,34 +13,39 @@ const TITLES = [
 class Title extends Component {
   constructor() {
     super();
-    this.state = {titleIndex:0 , fadeIn: true};
+    this.state = {titleIndex: 0, fadeIn: true};
   }
-  componentDidMount(){
-    console.log("inside Did Mount")
-    this.animateTitles();
-    this.timeout = setTimeout(() =>{
-      this.setState({fadeIn:false})
-    } , 2000);
-  }
-componentWillUnmount(){
-  clearInterval(this.titleInterval);
-  clearTimeout(this.timeout);
-}
 
-  animateTitles (){
-    this.titleInterval = setInterval(()=>{
-    const titleIndex = (this.state.titleIndex + 1) % TITLES.length
-    this.setState({titleIndex , fadeIn: true });
-    this.timeout = setTimeout(() =>{
-      this.setState({fadeIn:false})
-    } , 2000);
-    },4000);
-    
+  componentDidMount() {
+    this.animateTitles();
+    this.timeout = setTimeout(() => {
+      this.setState({fadeIn: false})
+    }, 2000);
   }
+
+  componentWillUnmount() {
+    clearInterval(this.titleInterval);
+    clearTimeout(this.timeout);
+  }
+
+  animateTitles() {
+    this.titleInterval = setInterval(() => {
+      const titleIndex = (this.state.titleIndex + 1) % TITLES.length;
+      this.setState({titleIndex, fadeIn: true});
+      this.timeout = setTimeout(() => {
+        this.setState({fadeIn: false})
+      }, 2000);
+    }, 4000);
+  }
+
   render() {
-     
-    return <p className= {this.state.fadeIn? "title-fade-in" : "title-fade-out"}>{TITLES[this.state.titleIndex]}</p>;
-   
+    return (
+      <div className="title-container">
+        <p className={`futuristic-title ${this.state.fadeIn ? "title-fade-in" : "title-fade-out"}`}>
+          {TITLES[this.state.titleIndex]}
+        </p>
+      </div>
+    );
   }
 }
 
