@@ -4,22 +4,16 @@ import linkedin from "../assest/linkedin.png"
 
 
 const ContactsData = [
+ 
     {
         id :1, 
-        link: "mailto:englishmansh70@gmail.com",
-        image: email,
-        target: "_blank",
-
-    },
-    {
-        id :2, 
         link: "https://github.com/MnsDew",
         image: github,
         target: "_blank",
 
     },
     {
-        id :3, 
+        id :2, 
         link: "https://www.linkedin.com/in/mg-mns-coding/",
         image: linkedin,
         target: "_blank",
